+++ astro.config.mjs
@@ -19,7 +19,9 @@ async function getHighlighter() {
 
 // https://astro.build/config
 export default defineConfig({
-  site: config.site.base_url ? config.site.base_url : "https://mejoresfinanzas.com",
+  site: config.site.base_url
+    ? config.site.base_url
+    : "https://mejoresfinanzas.com",
   base: config.site.base_path ? config.site.base_path : "/",
   trailingSlash: config.site.trailing_slash ? "always" : "never",
   server: {